requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings>=2.2.1
aiohttp==3.9.1
python-multipart==0.0.6

# Frontend
streamlit>=1.28.0
plotly>=5.17.0
pillow>=10.0.0

# AI / LLM
google-generativeai>=0.8.0
pinecone-client>=3.0.0

# ML / Data science
torch>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
joblib>=1.3.0
networkx>=3.2.0

# Data / storage
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
pdfminer.six>=20221105

# HuggingFace
datasets>=2.14.0
huggingface-hub>=0.19.0

# Utilities
requests>=2.31.0
watchdog>=6.0.0
tqdm>=4.66.0
